+++ Arm_Controller/potentiometer.h
@@ -5,7 +5,7 @@
 //function declarations 
 
 void potentiometerSetup(struct potentiometer_data * setup_data, int input_pin, int max_v, uint16_t scale_upper_bound, uint16_t scale_lower_bound);
-uint16_t readPotentiometer(struct potentiometer_data * input_data);
+int readPotentiometer(struct potentiometer_data * input_data);
 
 //structures
